+++ test/test_helper.exs
@@ -1,2 +1,2 @@
-ExUnit.configure(assert_receive_timeout: 500, exclude: [:http2, compression: :stress])
+ExUnit.configure(assert_receive_timeout: 500, exclude: [compression: :stress])
 ExUnit.start()